requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.33
botocore==1.27.33
Django==3.2
django-ckeditor==6.3.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-storages==1.12.3
jmespath==1.0.1
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
